[Update] refactor: move graph_viz to examples; tensor __repr__ use
