+++ nujo/math/scalar.py
@@ -4,8 +4,8 @@ from numpy import around as np_round
 from numpy import ceil as np_ceil
 from numpy import floor as np_floor
 
-from nujo.autodiff import Tensor
 from nujo.autodiff.functions import Power
+from nujo.autodiff.tensor import Tensor
 
 __all__ = [
     'sqrt',
